♻️ refactor(config): extract payload rule helpers
Move visual config payload rule parsing, comparison, validation, serialization, and option constants into a dedicated helper module.

Keep the useVisualConfig public exports stable while reducing the hook to YAML orchestration and reducer state.

Validated with type-check, config tests, lint, and Vite production build.
