+++ src/hooks/useVisualConfig.ts
@@ -1,15 +1,27 @@
 import { useCallback, useMemo, useReducer } from 'react';
 import { isMap, parse as parseYaml, parseDocument } from 'yaml';
 import type {
-  PayloadFilterRule,
-  PayloadParamEntry,
-  PayloadParamValueType,
-  PayloadRule,
   VisualConfigValues,
   VisualConfigValidationErrors,
-  PayloadParamValidationErrorCode,
 } from '@/types/visualConfig';
 import { DEFAULT_VISUAL_VALUES } from '@/types/visualConfig';
+import {
+  arePayloadFilterRulesEqual,
+  arePayloadRulesEqual,
+  hasPayloadParamValidationErrors,
+  parsePayloadFilterRules,
+  parsePayloadRules,
+  parseRawPayloadRules,
+  serializePayloadFilterRulesForYaml,
+  serializePayloadRulesForYaml,
+  serializeRawPayloadRulesForYaml,
+} from './visualConfigPayloadRules';
+
+export {
+  getPayloadParamValidationError,
+  VISUAL_CONFIG_PAYLOAD_VALUE_TYPE_OPTIONS,
+  VISUAL_CONFIG_PROTOCOL_OPTIONS,
+} from './visualConfigPayloadRules';
 
 function asRecord(value: unknown): Record<string, unknown> | null {
   if (value === null || typeof value !== 'object' || Array.isArray(value)) return null;
@@ -182,151 +194,6 @@ export function getVisualConfigValidationErrors(
   };
 }
 
-export function getPayloadParamValidationError(
-  param: PayloadParamEntry
-): PayloadParamValidationErrorCode | undefined {
-  const trimmedValue = param.value.trim();
-
-  switch (param.valueType) {
-    case 'number': {
-      if (!trimmedValue) return 'payload_invalid_number';
-      const parsed = Number(trimmedValue);
-      return Number.isFinite(parsed) ? undefined : 'payload_invalid_number';
-    }
-    case 'boolean': {
-      const normalized = trimmedValue.toLowerCase();
-      return normalized === 'true' || normalized === 'false'
-        ? undefined
-        : 'payload_invalid_boolean';
-    }
-    case 'json': {
-      if (!trimmedValue) return 'payload_invalid_json';
-      try {
-        JSON.parse(param.value);
-        return undefined;
-      } catch {
-        return 'payload_invalid_json';
-      }
-    }
-    default:
-      return undefined;
-  }
-}
-
-function hasPayloadParamValidationErrors(rules: PayloadRule[]): boolean {
-  return rules.some((rule) =>
-    rule.params.some((param) => Boolean(getPayloadParamValidationError(param)))
-  );
-}
-
-function deepClone<T>(value: T): T {
-  if (typeof structuredClone === 'function') return structuredClone(value);
-  return JSON.parse(JSON.stringify(value)) as T;
-}
-
-function arePayloadModelEntriesEqual(
-  left: PayloadRule['models'],
-  right: PayloadRule['models']
-): boolean {
-  if (left === right) return true;
-  if (left.length !== right.length) return false;
-  for (let i = 0; i < left.length; i += 1) {
-    const a = left[i];
-    const b = right[i];
-    if (!a || !b) return false;
-    if (a.id !== b.id || a.name !== b.name || a.protocol !== b.protocol) return false;
-  }
-  return true;
-}
-
-function arePayloadParamEntriesEqual(
-  left: PayloadRule['params'],
-  right: PayloadRule['params']
-): boolean {
-  if (left === right) return true;
-  if (left.length !== right.length) return false;
-  for (let i = 0; i < left.length; i += 1) {
-    const a = left[i];
-    const b = right[i];
-    if (!a || !b) return false;
-    if (a.id !== b.id || a.path !== b.path || a.valueType !== b.valueType || a.value !== b.value) {
-      return false;
-    }
-  }
-  return true;
-}
-
-function arePayloadRulesEqual(left: PayloadRule[], right: PayloadRule[]): boolean {
-  if (left === right) return true;
-  if (left.length !== right.length) return false;
-  for (let i = 0; i < left.length; i += 1) {
-    const a = left[i];
-    const b = right[i];
-    if (!a || !b) return false;
-    if (a.id !== b.id) return false;
-    if (!arePayloadModelEntriesEqual(a.models, b.models)) return false;
-    if (!arePayloadParamEntriesEqual(a.params, b.params)) return false;
-  }
-  return true;
-}
-
-function arePayloadFilterRulesEqual(
-  left: PayloadFilterRule[],
-  right: PayloadFilterRule[]
-): boolean {
-  if (left === right) return true;
-  if (left.length !== right.length) return false;
-  for (let i = 0; i < left.length; i += 1) {
-    const a = left[i];
-    const b = right[i];
-    if (!a || !b) return false;
-    if (a.id !== b.id) return false;
-    if (!arePayloadModelEntriesEqual(a.models, b.models)) return false;
-    if (a.params.length !== b.params.length) return false;
-    for (let j = 0; j < a.params.length; j += 1) {
-      if (a.params[j] !== b.params[j]) return false;
-    }
-  }
-  return true;
-}
-
-function parsePayloadParamValue(raw: unknown): { valueType: PayloadParamValueType; value: string } {
-  if (typeof raw === 'number') {
-    return { valueType: 'number', value: String(raw) };
-  }
-
-  if (typeof raw === 'boolean') {
-    return { valueType: 'boolean', value: String(raw) };
-  }
-
-  if (raw === null || typeof raw === 'object') {
-    try {
-      const json = JSON.stringify(raw, null, 2);
-      return { valueType: 'json', value: json ?? 'null' };
-    } catch {
-      return { valueType: 'json', value: String(raw) };
-    }
-  }
-
-  return { valueType: 'string', value: String(raw ?? '') };
-}
-
-function parseRawPayloadParamValue(raw: unknown): string {
-  if (typeof raw === 'string') return raw;
-
-  try {
-    const json = JSON.stringify(raw, null, 2);
-    return json ?? '';
-  } catch {
-    return String(raw ?? '');
-  }
-}
-
-function parsePayloadProtocol(raw: unknown): string | undefined {
-  if (typeof raw !== 'string') return undefined;
-  return raw.trim() ? raw : undefined;
-}
-
 function deleteLegacyApiKeysProvider(doc: YamlDocument): void {
   if (docHas(doc, ['auth', 'providers', 'config-api-key', 'api-key-entries'])) {
     doc.deleteIn(['auth', 'providers', 'config-api-key', 'api-key-entries']);
@@ -339,184 +206,9 @@ function deleteLegacyApiKeysProvider(doc: YamlDocument): void {
   deleteIfMapEmpty(doc, ['auth']);
 }
 
-function parsePayloadRules(rules: unknown): PayloadRule[] {
-  if (!Array.isArray(rules)) return [];
-
-  return rules.map((rule, index) => {
-    const record = asRecord(rule) ?? {};
-
-    const modelsRaw = record.models;
-    const models = Array.isArray(modelsRaw)
-      ? modelsRaw.map((model, modelIndex) => {
-          const modelRecord = asRecord(model);
-          const nameRaw =
-            typeof model === 'string' ? model : (modelRecord?.name ?? modelRecord?.id ?? '');
-          const name = typeof nameRaw === 'string' ? nameRaw : String(nameRaw ?? '');
-          return {
-            id: `model-${index}-${modelIndex}`,
-            name,
-            protocol: parsePayloadProtocol(modelRecord?.protocol),
-          };
-        })
-      : [];
-
-    const paramsRecord = asRecord(record.params);
-    const params = paramsRecord
-      ? Object.entries(paramsRecord).map(([path, value], pIndex) => {
-          const parsedValue = parsePayloadParamValue(value);
-          return {
-            id: `param-${index}-${pIndex}`,
-            path,
-            valueType: parsedValue.valueType,
-            value: parsedValue.value,
-          };
-        })
-      : [];
-
-    return { id: `payload-rule-${index}`, models, params };
-  });
-}
-
-function parsePayloadFilterRules(rules: unknown): PayloadFilterRule[] {
-  if (!Array.isArray(rules)) return [];
-
-  return rules.map((rule, index) => {
-    const record = asRecord(rule) ?? {};
-
-    const modelsRaw = record.models;
-    const models = Array.isArray(modelsRaw)
-      ? modelsRaw.map((model, modelIndex) => {
-          const modelRecord = asRecord(model);
-          const nameRaw =
-            typeof model === 'string' ? model : (modelRecord?.name ?? modelRecord?.id ?? '');
-          const name = typeof nameRaw === 'string' ? nameRaw : String(nameRaw ?? '');
-          return {
-            id: `filter-model-${index}-${modelIndex}`,
-            name,
-            protocol: parsePayloadProtocol(modelRecord?.protocol),
-          };
-        })
-      : [];
-
-    const paramsRaw = record.params;
-    const params = Array.isArray(paramsRaw) ? paramsRaw.map(String) : [];
-
-    return { id: `payload-filter-rule-${index}`, models, params };
-  });
-}
-
-function parseRawPayloadRules(rules: unknown): PayloadRule[] {
-  if (!Array.isArray(rules)) return [];
-
-  return rules.map((rule, index) => {
-    const record = asRecord(rule) ?? {};
-
-    const modelsRaw = record.models;
-    const models = Array.isArray(modelsRaw)
-      ? modelsRaw.map((model, modelIndex) => {
-          const modelRecord = asRecord(model);
-          const nameRaw =
-            typeof model === 'string' ? model : (modelRecord?.name ?? modelRecord?.id ?? '');
-          const name = typeof nameRaw === 'string' ? nameRaw : String(nameRaw ?? '');
-          return {
-            id: `raw-model-${index}-${modelIndex}`,
-            name,
-            protocol: parsePayloadProtocol(modelRecord?.protocol),
-          };
-        })
-      : [];
-
-    const paramsRecord = asRecord(record.params);
-    const params = paramsRecord
-      ? Object.entries(paramsRecord).map(([path, value], pIndex) => ({
-          id: `raw-param-${index}-${pIndex}`,
-          path,
-          valueType: 'json' as const,
-          value: parseRawPayloadParamValue(value),
-        }))
-      : [];
-
-    return { id: `payload-raw-rule-${index}`, models, params };
-  });
-}
-
-function serializePayloadRulesForYaml(rules: PayloadRule[]): Array<Record<string, unknown>> {
-  return rules
-    .map((rule) => {
-      const models = (rule.models || [])
-        .filter((m) => m.name?.trim())
-        .map((m) => {
-          const obj: Record<string, unknown> = { name: m.name.trim() };
-          if (m.protocol) obj.protocol = m.protocol;
-          return obj;
-        });
-
-      const params: Record<string, unknown> = {};
-      for (const param of rule.params || []) {
-        if (!param.path?.trim()) continue;
-        let value: unknown = param.value;
-        if (param.valueType === 'number') {
-          const num = Number(param.value);
-          value = Number.isFinite(num) ? num : param.value;
-        } else if (param.valueType === 'boolean') {
-          value = param.value === 'true';
-        } else if (param.valueType === 'json') {
-          try {
-            value = JSON.parse(param.value);
-          } catch {
-            value = param.value;
-          }
-        }
-        params[param.path.trim()] = value;
-      }
-
-      return { models, params };
-    })
-    .filter((rule) => rule.models.length > 0);
-}
-
-function serializePayloadFilterRulesForYaml(
-  rules: PayloadFilterRule[]
-): Array<Record<string, unknown>> {
-  return rules
-    .map((rule) => {
-      const models = (rule.models || [])
-        .filter((m) => m.name?.trim())
-        .map((m) => {
-          const obj: Record<string, unknown> = { name: m.name.trim() };
-          if (m.protocol) obj.protocol = m.protocol;
-          return obj;
-        });
-
-      const params = (Array.isArray(rule.params) ? rule.params : [])
-        .map((path) => String(path).trim())
-        .filter(Boolean);
-
-      return { models, params };
-    })
-    .filter((rule) => rule.models.length > 0);
-}
-
-function serializeRawPayloadRulesForYaml(rules: PayloadRule[]): Array<Record<string, unknown>> {
-  return rules
-    .map((rule) => {
-      const models = (rule.models || [])
-        .filter((m) => m.name?.trim())
-        .map((m) => {
-          const obj: Record<string, unknown> = { name: m.name.trim() };
-          if (m.protocol) obj.protocol = m.protocol;
-          return obj;
-        });
-
-      const params: Record<string, unknown> = {};
-      for (const param of rule.params || []) {
-        if (!param.path?.trim()) continue;
-        params[param.path.trim()] = param.value;
-      }
-
-      return { models, params };
-    })
-    .filter((rule) => rule.models.length > 0);
+function deepClone<T>(value: T): T {
+  if (typeof structuredClone === 'function') return structuredClone(value);
+  return JSON.parse(JSON.stringify(value)) as T;
 }
 
 type VisualConfigState = {
@@ -1156,68 +848,3 @@ export function useVisualConfig() {
     setVisualValues,
   };
 }
-
-export const VISUAL_CONFIG_PROTOCOL_OPTIONS = [
-  {
-    value: '',
-    labelKey: 'config_management.visual.payload_rules.provider_default',
-    defaultLabel: 'Default',
-  },
-  {
-    value: 'openai',
-    labelKey: 'config_management.visual.payload_rules.provider_openai',
-    defaultLabel: 'OpenAI',
-  },
-  {
-    value: 'openai-response',
-    labelKey: 'config_management.visual.payload_rules.provider_openai_response',
-    defaultLabel: 'OpenAI Response',
-  },
-  {
-    value: 'gemini',
-    labelKey: 'config_management.visual.payload_rules.provider_gemini',
-    defaultLabel: 'Gemini',
-  },
-  {
-    value: 'claude',
-    labelKey: 'config_management.visual.payload_rules.provider_claude',
-    defaultLabel: 'Claude',
-  },
-  {
-    value: 'codex',
-    labelKey: 'config_management.visual.payload_rules.provider_codex',
-    defaultLabel: 'Codex',
-  },
-  {
-    value: 'antigravity',
-    labelKey: 'config_management.visual.payload_rules.provider_antigravity',
-    defaultLabel: 'Antigravity',
-  },
-] as const;
-
-export const VISUAL_CONFIG_PAYLOAD_VALUE_TYPE_OPTIONS = [
-  {
-    value: 'string',
-    labelKey: 'config_management.visual.payload_rules.value_type_string',
-    defaultLabel: 'String',
-  },
-  {
-    value: 'number',
-    labelKey: 'config_management.visual.payload_rules.value_type_number',
-    defaultLabel: 'Number',
-  },
-  {
-    value: 'boolean',
-    labelKey: 'config_management.visual.payload_rules.value_type_boolean',
-    defaultLabel: 'Boolean',
-  },
-  {
-    value: 'json',
-    labelKey: 'config_management.visual.payload_rules.value_type_json',
-    defaultLabel: 'JSON',
-  },
-] as const satisfies ReadonlyArray<{
-  value: PayloadParamValueType;
-  labelKey: string;
-  defaultLabel: string;
-}>;
